22.01.08 회원기능 + msg 수정

+++ bbs/src/main/java/controller/DetailServlet.java
@@ -1,6 +1,7 @@
 package controller;
 
 import java.io.IOException;
+import java.io.PrintWriter;
 
 import javax.servlet.ServletException;
 import javax.servlet.annotation.WebServlet;
@@ -11,6 +12,7 @@ import javax.servlet.http.HttpSession;
 
 import model.dao.BoardDAO;
 import model.vo.BoardVO;
+//import model.vo.MembersVO;
 
 @WebServlet("/detail")
 public class DetailServlet extends HttpServlet {
@@ -25,42 +27,70 @@ public class DetailServlet extends HttpServlet {
 		
 		String boardPath = null;
 		
-		if (detail_btn == null) {
-
-			String num = request.getParameter("num");
-			BoardVO vo = dao.detailView(Integer.parseInt(num));
-			dao.cntIncrease(vo);
-			vo.setCnt(vo.getCnt()+1);
-			request.setAttribute("detailPage", vo);
-			boardPath = "/jsp/BoardDetail.jsp";
-			
-		} else if (detail_btn.equals("list")) {
-			boardPath = "/board";
-			
-		} else {
-			//BoardVO vo = (BoardVO)request.getAttribute("detail");
-			
-			HttpSession session = request.getSession();
-			BoardVO vo = (BoardVO)session.getAttribute("detail");
-			
-			if (detail_btn.equals("update")) {
-				request.setAttribute("updatePage", vo);
-				session.removeAttribute("detail");
-				boardPath = "/jsp/UpdateBoard.jsp";
+		HttpSession session = request.getSession();
+		
+		if(session.getAttribute("user") != null) {
+			if (detail_btn == null) {
+				String num = request.getParameter("num");
+				BoardVO vo = dao.detailView(Integer.parseInt(num));
+				dao.cntIncrease(vo);
+				vo.setCnt(vo.getCnt()+1);
+				request.setAttribute("detailPage", vo);
+				boardPath = "/jsp/BoardDetail.jsp";
 				
-			} else if (detail_btn.equals("delete")) {
-				dao.delete(vo.getNum());
-				session.removeAttribute("detail");
+			} else if (detail_btn.equals("list")) {
 				boardPath = "/board";
-			}
-			else if(detail_btn.equals("write")) {
+			} else {
+				//BoardVO vo = (BoardVO)request.getAttribute("detail");
 				
-				boardPath = "/jsp/WriteBoard.jsp";
+				BoardVO vo = (BoardVO)session.getAttribute("detail");
+				
+				if (detail_btn.equals("update")) {
+					request.setAttribute("updatePage", vo);
+					
+					boardPath = "/jsp/UpdateBoard.jsp";
+				} else if (detail_btn.equals("delete")) {
+					dao.delete(vo.getNum());
+					
+					boardPath = "/board";
+				}
+				else if(detail_btn.equals("write")) {
+					
+					boardPath = "/jsp/WriteBoard.jsp";
+				}
+				
+				session.removeAttribute("detail");
 			}
-			
+		}
+		else {
+			System.out.println("로그인을 해주세요");
+			boardPath = "/board";
 		}
 		
 		request.getRequestDispatcher(boardPath).forward(request, response);
+	}
+	
+	@Override
+	protected void doPost(HttpServletRequest request, HttpServletResponse response) throws ServletException, IOException {
+		response.setContentType("text/html; charset=UTF-8");
+		request.setCharacterEncoding("UTF-8");
+		String boardPath = "/bbs/board";
+		BoardDAO dao = new BoardDAO();
+		
+		// WriteBoard (writer, title, content)
+		PrintWriter out = response.getWriter();
+		String b_title = request.getParameter("b_title");
+		String b_writer = request.getParameter("b_writer");
+		String b_content = request.getParameter("b_content");
+		
+		BoardVO boardVO = new BoardVO();
+		boardVO.setTitle(b_title);
+		boardVO.setWriter(b_writer);
+		boardVO.setContent(b_content);
+		
+		dao.insert(boardVO);
+
+		response.sendRedirect(boardPath);
 		
 	}
 }
